Update Transaction Model
- Seperate Transaction SQLModel database and the Pydentic model.
- Create a API models module for the DTO
- Create DataBase models module for the SQLModel Database models.
- create Utils module for the TransactionType
