+++ src/expense_tracker_plus/main.py
@@ -1,14 +1,9 @@
-from fastapi import FastAPI, Request
-from .routers.transaction_router import transactions_router
+from fastapi import FastAPI
+from .api.routers.transaction_router import transactions_router
 from .data_access.database_manager import DatabaseManager
 
 
 app = FastAPI()
-
-
-def db_session(request: Request):
-    return DatabaseManager().create_session()
-
 
 app.include_router(transactions_router)
