+++ src/expense_tracker_plus/utils/transactions_type.py
@@ -1,0 +1,6 @@
+from enum import StrEnum
+
+
+class TransactionType(StrEnum):
+    Income = "Income"
+    Expense = "Expense"
